requirements.txt
aws-cdk-lib>=2.170.0
boto3>=1.35.0
constructs>=10.4.2
pypdf>=5.0.0
python-dotenv>=1.0.1
